Visual Studio solution, Test Map, Fixed Extra Functions
Just filled out a massive description and git hub crashed.
so mad. anyway.

1) Added a Visual Studio solution that edits the files directly. If you
want to use this solution you will have to go to project > Properties>
Configuration Properties > c/c++ > Precompiled Headers, and change the
"Precompiled Header" setting to "Not Using Precompiled Headers" option.
this means you will not have to include "stdafx.h" for it to run.

2) Added the base static test map to the map class, and a corresponding
print function. This is currently displayed directly underneath the
random map.

3) Changed the getch function , now getKey() , to work on both windows
and Linux, can be used void or the return can be assigned to a variable.
the clear_screen() function is also fixed.